refactor: ctpat urls
